Refactor Zappy AI implementation: - Consolidated AI logic into a single file (ai.py) and removed main.py. - Enhanced AI capabilities with neural networks, adaptive learning, and advanced decision-making. - Introduced performance metrics tracking and experience-based learning. - Removed deprecated methods and streamlined action execution. - Updated connection handling and response processing for improved reliability. - Added model persistence for saving and loading AI state.
