+++ ai/src/NeuralNetwork.py
@@ -4,6 +4,16 @@
 ## Made by adrien.marette@epitech.eu
 ## Login   <adrien.marette@epitech.eu>
 ##
+## Started on  Tue Jun 10 2:38:29 PM 2025 adrien.marette@epitech.eu
+## Last update Wed Jun 10 2:38:32 PM 2025 adrien.marette@epitech.eu
+##
+
+##
+## NeuralNetwork.py for Zappy in /home/adrien/Epitech/B-YEP-410/Zappy/ai/src
+##
+## Made by adrien.marette@epitech.eu
+## Login   <adrien.marette@epitech.eu>
+##
 ## Started on  Tue Jun 10 1:12:56 PM 2025 adrien.marette@epitech.eu
 ## Last update Wed Jun 10 1:44:10 PM 2025 adrien.marette@epitech.eu
 ##
@@ -14,7 +24,7 @@ import numpy as np
 from typing import List
 
 class NeuralNetwork:
-    """Simple feedforward neural network with backprogramation"""
+    """Simple feedforward neural network with backpropagation"""
     
     def __init__(self, input_size: int, hidden_sizes: List[int], output_size: int, learning_rate: float = 0.001):
         self.learning_rate = learning_rate
@@ -30,7 +40,7 @@ class NeuralNetwork:
     def sigmoid(self, x):
         return 1 / (1 + np.exp(-np.clip(x, -500, 500)))
     
-    def sigmoid_relative(self, x):
+    def sigmoid_derivative(self, x):
         return x * (1 - x)
     
     def relu(self, x):
@@ -45,9 +55,9 @@ class NeuralNetwork:
     
         for i, layer in enumerate(self.layers):
             z = np.dot(current_input, layer['weight']) + layer['bias']
-            if i == len(self.layers) - 1: # Output layer
+            if i == len(self.layers) - 1:  # Output layer
                 activation = self.sigmoid(z)
-            else: # Hidden layers
+            else:  # Hidden layers
                 activation = self.relu(z)
 
             self.activations.append(activation)
@@ -60,14 +70,17 @@ class NeuralNetwork:
 
         # Calculate output error
         output_error = y_pred - y_true
-        deltas = [output_error]
+        deltas = []
 
-        # Backpropagate errors
-        for i in range(len(self.layers) - 1, 0, -1):
-            if i == len(self.layers) - 1: 
-                delta = output_error * self.sigmoid_derivative(self.activations[i + 1])
-            else:
-                delta = deltas[0].dot(self.layers[i + 1]['weigth'].T) * self.relu_derivative(self.activations[i + 1])
+        # Calculate delta for output layer
+        output_delta = output_error * self.sigmoid_derivative(y_pred)
+        deltas.append(output_delta)
+
+        # Backpropagate errors through hidden layers
+        for i in range(len(self.layers) - 2, -1, -1):
+            # Calculate delta for current layer
+            error = deltas[0].dot(self.layers[i + 1]['weight'].T)
+            delta = error * self.relu_derivative(self.activations[i + 1])
             deltas.insert(0, delta)
 
         # Update weights and biases
@@ -85,11 +98,24 @@ class NeuralNetwork:
     
     def save(self, filepath):
         with open(filepath, 'wb') as f:
-            pickle.dump(self.layers, f)
+            pickle.dump({
+                'layers': self.layers,
+                'learning_rate': self.learning_rate
+            }, f)
 
     def load(self, filepath):
         if os.path.exists(filepath):
-            with open(filepath, 'rb') as f:
-                self.layers = pickle.load(f)
-            return True
+            try:
+                with open(filepath, 'rb') as f:
+                    data = pickle.load(f)
+                    if isinstance(data, dict):
+                        self.layers = data['layers']
+                        self.learning_rate = data.get('learning_rate', self.learning_rate)
+                    else:
+                        # Handle old format
+                        self.layers = data
+                return True
+            except Exception as e:
+                print(f"Warning: Could not load model from {filepath}: {e}")
+                return False
         return False
